6_Find_Similar_Thing_App_Kids/SimilarWordsFinderApp/requirements.txt
langchain
streamlit
openai
tiktoken
python-dotenv
faiss-cpu
langchain-community
openpyxl